⚡️ fix : 답변란의 텍스트 입력 방법 개선
